+++ R/app.r
@@ -42,12 +42,11 @@
 #'
 #' @returns A `GRaster`.
 #'
-#' @seealso [terra::app()], [terra::lapp()], [subst()], [classify()], and especially **GRASS** manual page for module `r.mapcalc` (see `grassHelp("r.mapcalc")`)
+#' @seealso [terra::app()], [terra::lapp()], [subst()], [classify()], and especially the **GRASS** manual page for module `r.mapcalc` (see `grassHelp("r.mapcalc")`)
 #'
 #' @example man/examples/ex_app.r
 #'
-#' @name app
-#' @aliases app,lapp
+#' @aliases app
 #' @rdname app
 #' @exportMethod app
 methods::setMethod(