+++ R/project.r
@@ -30,7 +30,7 @@
 #' * Two numeric values: Values for the new resolution (x- and y-dimensions).
 #' * `"center"`: This method locates the centroid of the raster to be projected (in the same CRS as the original raster). It then creates four points north, south, east, and west of the centroid, each spaced one cell's width from the centroid. This set of points is then projected to the new CRS. The new cell size in the x-dimension will be the average of the distance between the east and west points from the centroid, and in the y-dimension the average from the centroid to the north and south points.
 #' * `"fallback"` (default): This applies the `terra` method first, but if that fails, then tries `template`, then `center`. This process can take a long time for large rasters.
-#' 
+#'
 #' @param wrap Logical:
 #' * `GRaster`s: When projecting rasters that "wrap around" (i.e., whole-world rasters or rasters that have edges that actually circle around to meet on the globe), `wrap` should be `TRUE` to avoid removing rows and columns from the "edge" of the map. The default is `FALSE`.
 #' * `GVector`s: When projecting vectors that span the international date line at 180E/W, `wrap` should be `TRUE` to avoid an issue where the coordinates are incorrectly mapped to the range -180 to 180.
@@ -83,19 +83,22 @@ methods::setMethod(
 	} else if (res == "fallback") {
 	
 		out <- FALSE
-		j <- 0L
-		n <- length(res)
+		j <- 1L
 		try <- c("terra", "template", "center")
-		while (inherits(out, "logical") & j <= n) {
+		n <- length(try)
+		while (is.logical(out) & j <= n) {
 		
-			j <- j + 1L
 			thisRes <- try[j]
 
 			out <- tryCatch(
 				.projectRaster(x = x, y = y, align = align, method = method, fallback = fallback, res = thisRes, wrap = wrap, verbose = verbose),
 				error = function(cond) FALSE
 			)
+
+			if (verbose & is.logical(out)) omnibus::say("The ", try, " `res` method failed. Trying method `", try[j + 1L], "`.")
 		
+			j <- j + 1L
+
 		}
 	
 	}
@@ -122,10 +125,11 @@ methods::setMethod(
 	# imperfect catch for cases where `wrap` should be `TRUE` but is not
 	unproj <- .projection(x) %in%
 		c("Latitude-Longitude", "WGS84", "WGS 84", "NAD83", "NAD 83", "NAD27", "NAD 27")
+
 	if (unproj & !wrap) {
 
 		extent <- ext(x, vector = TRUE)
-		if ((extent[1L] == -180 & extent[2L] == 180) | (extent[3L] == -90 & extent[4L] == 90)) warning("This GRaster seems to wrap around the globe to meet at the international\n  date line and/or the poles. Should `wrap` be `TRUE`?")
+		if ((extent[1L] == -180 & extent[2L] == 180) | (extent[3L] == -90 & extent[4L] == 90)) warning("This GRaster seems to wrap around the globe to meet at the international\n  date line and/or the poles. Should `wrap` be TRUE?")
 
 	}
 	
@@ -136,12 +140,14 @@ methods::setMethod(
 	
 		if (yLocation == xLocation) {
 			warning("Object is already in the desired coordinate reference system.")
-			return()
+			return(x)
 		}
 	
 	} else if (is.null(yLocation)) {
+	
 		yLocation <- .locationCreate(y)
 		yLocation <- .location(yLocation)
+	
 	}
 	
 	# method
@@ -161,7 +167,7 @@ methods::setMethod(
 	}
 
 	nSteps <- nLayers + (res == "terra")
-	if (verbose | faster("verbose")) {
+	if (verbose & nLayers > 1L) {
 		pb <- utils::txtProgressBar(min = 0, max = nSteps, initial = 0, style = 3, width = 30)
 		steps <- 0
 	}
@@ -177,26 +183,24 @@ methods::setMethod(
 		# resample as per terra::project()
 		} else if (res == "terra") {
 
-			if (verbose | faster("verbose")) {
+			if (verbose & nLayers > 1L) {
 				utils::setTxtProgressBar(pb, steps)
 				steps <- steps + 1
 			}
 
 			# Use a SpatRaster as template for resampling region. We do this so we can set the "region" resolution and extent correctly.
 			extent <- ext(x, vector = TRUE)
-			xRast <- matrix(NA_real_, nrow = nrow(x), ncol = ncol(x))
-			xRast <- terra::rast(xRast, crs = crs(x), extent = extent)
+			xRast <- terra::rast(nrows = nrow(x), ncols = ncol(x), crs = crs(x), extent = extent)
 			
 			extent <- ext(y, vector = TRUE)
-			yRast <- matrix(NA_real_, nrow = nrow(y), ncol = ncol(y))
-			yRast <- terra::rast(yRast, crs = crs(y), extent = extent)
+			yRast <- terra::rast(nrows = nrow(y), ncols = ncol(y), crs = crs(y), extent = extent)
 
-			xRast <- terra::project(xRast, yRast, align = TRUE)
-			xRast <- terra::project(xRast, crs(x))
+			xRast <- terra::project(xRast, yRast, align = align, threads = faster("cores") > 1) # resample/project
+			xRast <- terra::project(xRast, crs(x), threads = faster("cores") > 1) # project back to start
 			
 			xSR <- xRast
 
-			xRast[] <- 1L
+			xRast <- terra::init(xRast, fun = "chess")
 			xRast <- fast(xRast)
 
 			# resample x in its native location to the resolution it will have in the target location
@@ -207,10 +211,13 @@ methods::setMethod(
 				fallback = fallback
 			)
 
-			xRast <- terra::project(xSR, yRast, method = "near", align = align)
+			xRast <- terra::project(xSR, yRast, method = "near", align = align, threads = faster("cores") > 1)
 
 			# reshape region
+
+	omnibus::say('&&&&&&& start at .locationRestore in project()')
 			.locationRestore(yLocation)
+	omnibus::say('&&&&&&& stop at .locationRestore in project()')
 			.region(xRast)
 
 		# } else if (res == "dimensions") {
@@ -329,7 +336,7 @@ methods::setMethod(
 			xRast <- matrix(NA_integer_, nrow = nrow(x), ncol = ncol(x))
 			xSR <- terra::rast(xRast, crs = crs(x), extent = extent)
 			
-			xRast <- terra::project(xSR, crs(y), method = "near", align = align)
+			xRast <- terra::project(xSR, crs(y), method = "near", align = align, threads = faster("cores") > 1)
 
 			# reshape region
 			.locationRestore(yLocation)
@@ -355,7 +362,7 @@ methods::setMethod(
 	srcs <- .makeSourceName("project", "raster", nlyr(x))
 	for (i in seq_len(nlyr(x))) {
 		
-		if (verbose | faster("verbose")) {
+		if (verbose & nLayers > 1L) {
 			utils::setTxtProgressBar(pb, steps)
 			steps <- steps + 1
 		}
@@ -369,11 +376,18 @@ methods::setMethod(
 			memory = faster("memory"),
 			flags = c(.quiet(), "overwrite")
 		)
-		args <- .addLocationProject(args, .location(x))
+
+		ver <- grassInfo("versionNumber")
+		if (ver <= 8.3) {
+			args$location <- .location(x)
+		} else {
+			args$project <- .location(x)
+		}
 
 		if (wrap) args$flags <- c(args$flags, "n")
 
-		do.call(rgrass::execGRASS, args=args)
+		do.call(rgrass::execGRASS, args = args)
+
 		if (is.factor(x)[i] & method == "nearest") {
 			levels <- cats(x)[[i]]
 		} else {
@@ -388,11 +402,13 @@ methods::setMethod(
 
 	} # project next raster
 
-	.locationRestore(y)
+	omnibus::say('4444444444 start locationRestore out')
+	.locationRestore(out)
+	omnibus::say('4444444444 stop locationRestore out')
 
 	# if using y as extent to which to crop
 	if (!align & inherits(y, "GRaster")) out <- crop(out, y)
-	if (verbose | faster("verbose")) close(pb)
+	if (verbose & nLayers > 1) close(pb)
 	out
 
 }
@@ -452,7 +468,13 @@ methods::setMethod(
 		output = src,
 		flags = c(.quiet(), "overwrite")
 	)
-	args <- .addLocationProject(args, .location(x))
+
+	ver <- grassInfo("versionNumber")
+	if (ver <= 8.3) {
+		args$location <- .location(x)
+	} else {
+		args$project <- .location(x)
+	}
  
 	if (wrap) args$flags <- c(args$flags, "w") # if crosses international date line
 	# if (geomtype(x, grass = TRUE) == "point") args$flags <- c(args$flags, "b") # disable topology build for points