+++ R/spatSample.r
@@ -107,7 +107,6 @@ methods::setMethod(
 		ys[ys < extent[3L]] <- extent[3L]
 		ys[ys > extent[4L]] <- extent[4L]
 
-
 		### use v.in.ascii to ingest each subset of points
 		nAtATime <- 2E5 # optimal-ish size based on manual checks
 		sets <- ceiling(size / nAtATime)
@@ -196,7 +195,7 @@ methods::setMethod(
 			
 		}
 
-		if (faster("clean")) .rm(srcs, type = "vector", warn = FALSE)
+		.rm(srcs, type = "vector", warn = FALSE)
 		if ((verbose | faster("verbose")) & sets > 1)  close(pb)
 
 	# if strata is not NULL