+++ man/examples/ex_options.r
@@ -2,13 +2,9 @@ if (grassStarted()) {
 
 # See current values for options:
 faster("grassDir")
-faster("addonsDir")
-faster("clean")
 faster("cores")
 faster("memory")
 faster("useDataTable")
-faster("verbose")
-faster("workDir")
 faster() # all options
 
 # See default values for options:
@@ -18,14 +14,14 @@ faster(default = TRUE) # all options
 # Set options (change accordingly for your system!!!)
 if (FALSE) {
 
-opts. <- faster() # remember starting values of options
+   opts. <- faster() # remember starting values of options
 
-faster(grassDir = "C:/Program Files/GRASS GIS 8.4")
-faster(verbose = TRUE, memory = 1024, cores = 1)
+   faster(grassDir = "C:/Program Files/GRASS GIS 8.4")
+   faster(verbose = TRUE, memory = 1024, cores = 1)
 
-faster(c("grassDir", "verbose", "memory", "cores"))
+   faster(c("grassDir", "verbose", "memory", "cores"))
 
-faster(opts.) # reset options to starting values
+   faster(opts.) # reset options to starting values
 
 }